UI re-org (#14)
* Fixing on homepage

* re-org on UI structure

* UI further update and clearing all warnings

* minor update

* adding UI wireframe
