+++ ui/src/App.js
@@ -1,37 +1,25 @@
 import React from 'react';
-import Homepage from './Homepage';
-import 'semantic-ui-css/semantic.min.css';
-import { SubstrateContextProvider } from './substrate-lib';
-import ChainData from './ChainData';
 import {
   BrowserRouter as Router,
   Switch,
   Route
 } from 'react-router-dom';
-import ShipmentTrackingPage from './ShipmentTrackingPage';
 
-function Main () {
+import Dashboard from './Dashboard';
+import Homepage from './Homepage';
+import 'semantic-ui-css/semantic.min.css';
+
+export default function App () {
   return (
     <Router>
       <Switch>
         <Route path="/demo">
-          <ChainData />
-        </Route>
-        <Route path="/ShipmentTracking">
-          <ShipmentTrackingPage />
+          <Dashboard />
         </Route>
         <Route path="/">
           <Homepage/>
         </Route>
       </Switch>
     </Router>
-  );
-}
-
-export default function App () {
-  return (
-    <SubstrateContextProvider>
-      <Main />
-    </SubstrateContextProvider>
   );
 }
