+++ ui/src/Homepage.js
@@ -5,10 +5,10 @@ import {
   Button,
   Container,
   Divider,
+  Dropdown,
   Grid,
   Header,
   Icon,
-  List,
   Menu,
   Responsive,
   Segment,
@@ -89,12 +89,14 @@ class DesktopContainer extends Component {
               size='large'
             >
               <Container>
-                <Menu.Item as='a' active>
-                  Home
-                </Menu.Item>
+                <Menu.Item as='a' active>Home</Menu.Item>
                 <Menu.Item as={Link} to='/demo'>Demo</Menu.Item>
-                <Menu.Item href="https://substrate.io">Substrate</Menu.Item>
-                <Menu.Item href="https://parity.io">Parity</Menu.Item>
+                <Dropdown item text = 'External'>
+                  <Dropdown.Menu>
+                    <Dropdown.Item href="https://substrate.dev" target="_blank">Substrate</Dropdown.Item>
+                    <Dropdown.Item href="https://parity.io" target="_blank">Parity</Dropdown.Item>
+                  </Dropdown.Menu>
+                </Dropdown>
               </Container>
             </Menu>
             <HomepageHeading/>
@@ -136,11 +138,16 @@ class MobileContainer extends Component {
           vertical
           visible={sidebarOpened}
         >
-          <Menu.Item as='a' active>
-            Home
+          <Menu.Item as='a' active>Home</Menu.Item>
+          <Menu.Item as={Link} to='/demo'>Demo</Menu.Item>
+          <Menu.Item as='a' href="https://substrate.dev" target="_blank">
+            Substrate
+            <Icon name='external' />
           </Menu.Item>
-          <Menu.Item href="https://substrate.io">Substrate</Menu.Item>
-          <Menu.Item href="https://parity.io">Parity</Menu.Item>
+          <Menu.Item as='a' href="https://parity.io" target="_blank">
+            Parity
+            <Icon name='external' />
+          </Menu.Item>
         </Sidebar>
 
         <Sidebar.Pusher dimmed={sidebarOpened}>
@@ -189,117 +196,45 @@ const Homepage = () => (
         <Grid.Row>
           <Grid.Column width={8}>
             <Header as='h3' style={{ fontSize: '2em' }}>
-              We Help Companies and Companions
+              What is this demo about?
             </Header>
             <p style={{ fontSize: '1.33em' }}>
-              We can give your company superpowers to do things that they never thought possible.
-              Let us delight your customers and empower your needs... through pure data analytics.
-            </p>
-            <Header as='h3' style={{ fontSize: '2em' }}>
-              We Make Bananas That Can Dance
-            </Header>
-            <p style={{ fontSize: '1.33em' }}>
-              Yes that's right, you thought it was the stuff of dreams, but even bananas can be
-              bioengineered.
+              TODO TODO TODO
             </p>
           </Grid.Column>
         </Grid.Row>
-        <Grid.Row>
-          <Grid.Column textAlign='center'>
-            <Button size='huge'>Check Them Out</Button>
-          </Grid.Column>
-        </Grid.Row>
-      </Grid>
-    </Segment>
 
-    <Segment style={{ padding: '0em' }} vertical>
-      <Grid celled='internally' columns='equal' stackable>
-        <Grid.Row textAlign='center'>
-          <Grid.Column style={{ paddingBottom: '5em', paddingTop: '5em' }}>
-            <Header as='h3' style={{ fontSize: '2em' }}>
-              "What a Company"
-            </Header>
-            <p style={{ fontSize: '1.33em' }}>That is what they all say about us</p>
-          </Grid.Column>
-          <Grid.Column style={{ paddingBottom: '5em', paddingTop: '5em' }}>
-            <Header as='h3' style={{ fontSize: '2em' }}>
-              "I shouldn't have gone with their competitor."
-            </Header>
-          </Grid.Column>
-        </Grid.Row>
-      </Grid>
-    </Segment>
-
-    <Segment style={{ padding: '8em 0em' }} vertical>
-      <Container text>
-        <Header as='h3' style={{ fontSize: '2em' }}>
-          Breaking The Grid, Grabs Your Attention
-        </Header>
-        <p style={{ fontSize: '1.33em' }}>
-          Instead of focusing on content creation and hard work, we have learned how to master the
-          art of doing nothing by providing massive amounts of whitespace and generic content that
-          can seem massive, monolithic and worth your attention.
-        </p>
-        <Button as='a' size='large'>
-          Read More
-        </Button>
-
-        <Divider
-          as='h4'
-          className='header'
-          horizontal
+        <Divider as='h4' className='header' horizontal
           style={{ margin: '3em 0em', textTransform: 'uppercase' }}
-        >
-          {/* <a href='#'>Case Studies</a> */}
-        </Divider>
+        />
 
-        <Header as='h3' style={{ fontSize: '2em' }}>
-          Did We Tell You About Our Bananas?
-        </Header>
-        <p style={{ fontSize: '1.33em' }}>
-          Yes I know you probably disregarded the earlier boasts as non-sequitur filler content, but
-          it's really true. It took years of gene splicing and combinatory DNA research, but our
-          bananas can really dance.
-        </p>
-        <Button as='a' size='large'>
-          I'm Still Quite Interested
-        </Button>
-      </Container>
-    </Segment>
+        <Grid.Row>
+          <Grid.Column width={8}>
+            <Header as='h3' style={{ fontSize: '2em' }}>
+              Related Github Repositories
+            </Header>
+            <p style={{ fontSize: '1.33em' }}>
+              TODO TODO TODO
+            </p>
+          </Grid.Column>
+        </Grid.Row>
 
-    <Segment inverted vertical style={{ padding: '5em 0em' }}>
-      <Container>
-        <Grid divided inverted stackable>
-          <Grid.Row>
-            <Grid.Column width={3}>
-              <Header inverted as='h4' content='About' />
-              <List link inverted>
-                <List.Item as='a'>Sitemap</List.Item>
-                <List.Item as='a'>Contact Us</List.Item>
-                <List.Item as='a'>Religious Ceremonies</List.Item>
-                <List.Item as='a'>Gazebo Plans</List.Item>
-              </List>
-            </Grid.Column>
-            <Grid.Column width={3}>
-              <Header inverted as='h4' content='Services' />
-              <List link inverted>
-                <List.Item as='a'>Banana Pre-Order</List.Item>
-                <List.Item as='a'>DNA FAQ</List.Item>
-                <List.Item as='a'>How To Access</List.Item>
-                <List.Item as='a'>Favorite X-Men</List.Item>
-              </List>
-            </Grid.Column>
-            <Grid.Column width={7}>
-              <Header as='h4' inverted>
-                Footer Header
-              </Header>
-              <p>
-                Extra space for a call to action inside the footer that could help re-engage users.
-              </p>
-            </Grid.Column>
-          </Grid.Row>
-        </Grid>
-      </Container>
+        <Divider as='h4' className='header' horizontal
+          style={{ margin: '3em 0em', textTransform: 'uppercase' }}
+        />
+
+        <Grid.Row>
+          <Grid.Column width={8}>
+            <Header as='h3' style={{ fontSize: '2em' }}>
+              Contact Us
+            </Header>
+            <p style={{ fontSize: '1.33em' }}>
+              TODO TODO TODO
+            </p>
+          </Grid.Column>
+        </Grid.Row>
+
+      </Grid>
     </Segment>
   </ResponsiveContainer>
 );
