--- ui/src/Transfer.js
@@ -1,65 +1,0 @@
-import React, { useState } from 'react';
-import { Form, Input, Grid, Label, Icon } from 'semantic-ui-react';
-import { TxButton } from './substrate-lib/components';
-
-export default function Main (props) {
-  const [status, setStatus] = useState(null);
-  const [formState, setFormState] = useState({ addressTo: null, amount: 0 });
-  const { accountPair } = props;
-
-  const onChange = (_, data) =>
-    setFormState(prev => ({ ...prev, [data.state]: data.value }));
-
-  const { addressTo, amount } = formState;
-
-  return (
-    <Grid.Column width={8}>
-      <h1>Transfer</h1>
-      <Form>
-        <Form.Field>
-          <Label
-            basic
-            color='teal'
-          >
-            <Icon name='hand point right' />
-            1 Unit = 1000000000000
-          </Label>
-        </Form.Field>
-        <Form.Field>
-          <Input
-            fluid
-            label='To'
-            type='text'
-            placeholder='address'
-            state='addressTo'
-            onChange={onChange}
-          />
-        </Form.Field>
-        <Form.Field>
-          <Input
-            fluid
-            label='Amount'
-            type='number'
-            state='amount'
-            onChange={onChange}
-          />
-        </Form.Field>
-        <Form.Field style={{ textAlign: 'center' }}>
-          <TxButton
-            accountPair={accountPair}
-            label='Submit'
-            type='SIGNED-TX'
-            setStatus={setStatus}
-            attrs={{
-              palletRpc: 'balances',
-              callable: 'transfer',
-              inputParams: [addressTo, amount],
-              paramFields: [true, true]
-            }}
-          />
-        </Form.Field>
-        <div style={{ overflowWrap: 'break-word' }}>{status}</div>
-      </Form>
-    </Grid.Column>
-  );
-}
