+++ ui/src/components/Members.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function Main (props) {
+  return (
+    <h1>Members</h1>
+  );
+}
