+++ ui/src/components/index.js
@@ -1,0 +1,14 @@
+import AccountSelector from './AccountSelector';
+import BlockNumber from './BlockNumber';
+import Events from './Events';
+import Members from './Members';
+import NodeInfo from './NodeInfo';
+import Organizations from './Organizations';
+import Products from './Products';
+import Shipments from './Shipments';
+import TopNavMenu from './TopNavMenu';
+
+export {
+  AccountSelector, BlockNumber, Events, Members, NodeInfo, Organizations, Products, Shipments,
+  TopNavMenu
+};
